Release v0.11 for BenchmarkSVs; check Roadmap for CHANGELOG
