+++ BenchmarkSVs/SVisualizer/user_config.py
@@ -1,0 +1,9 @@
+COVARIATE_X = "Experiment"
+SKIP_COMP_HWE = True
+MAKE_MISSING_PASS_FILTER = True
+EXPERIMENT_ORDER = ['GATK-SV_HGSVC2', 'Sniffles2x_vs_HGSVC2', 'Sniffles5x_vs_HGSVC2', 'Sniffles8x_vs_HGSVC2']
+TRUVARI_DUP_TO_INS = True
+
+INCLUDE_QC = False
+INCLUDE_WITTYER = False
+INCLUDE_TRUVARI = True
